requirements.txt

streamlit
matplotlib
rdkit-pypi
pandas
numpy
pycaret
streamlit_ketcher
py3Dmol
stmol
ipython_genutils
pubchempy